requirements.txt
# eBay Store Automation Tool
# Python dependencies

# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1

# eBay API
ebaysdk==2.2.0

# Scheduling
APScheduler==3.10.4

# Database
SQLAlchemy==2.0.23

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# Logging
colorlog==6.8.0

# WSGI server for production
gunicorn==21.2.0

# Poshmark scraping dependencies
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
Pillow==10.1.0
